Complete ComboBox accessibility refactor
Replace old ComboBox implementation with improved ComboBoxAlt version,
keeping the ComboBox file name. This completes the accessibility
improvements for screen reader support.

Key changes:
- Replace old ComboBox with ComboBoxAlt implementation, keeping
  ComboBox.tsx file name for compatibility
- Fix race condition where getSelections() returned stale data when
  called from onUpdateSelection callback by adding selectedMapRef
- Add aria-live wrapper to PillBox to prevent screen reader
  announcements on selection changes
- Remove external clear buttons from trustee filters, use internal
  ComboBox clear button with absolute positioning
- Prevent Enter key from submitting forms when expanding ComboBox
- Update all test selectors to match new ComboBox structure
- Add UTM networking documentation for screen reader testing

Jira ticket: N/A
