+++ user-interface/src/trustees/filters/TrusteeDistrictFilter.scss
@@ -88,14 +88,15 @@
           white-space: nowrap;
           border: 0;
         }
-      }
 
-      // Position clear button absolutely in top right
-      > .filter-clear-button-container {
-        position: absolute;
-        top: 0;
-        right: 0;
-        z-index: 1;
+        .clear-all-button {
+          position: absolute;
+          top: -24px;
+          right: 0;
+          z-index: 1;
+          line-height: 1;
+          font-size: 16px;
+        }
       }
     }
